Tutorials/langgraph/requirements.txt
# LangGraph 학습 튜토리얼 필수 패키지

# Google Gemini API
google-generativeai>=0.3.0

# LangChain 및 LangGraph
langchain>=0.1.0
langchain-google-genai>=0.0.11
langgraph>=0.0.20

# 데이터 모델링
pydantic>=2.0.0

# 환경변수 관리
python-dotenv>=1.0.0

# 유틸리티
typing-extensions>=4.5.0
